Opponent is now better by default. It plays a winning move if it can, and it blocks the model if it has the ability to.
Episode generation is considerably slower, since you need to have logic which checks if the opponent actually has a winning move, and if not, you need to check if the opponent has a legal blocking move.

And in the end, if something about the DQN algorithm is not changed (we will probably need a mix of opponents instead of just 1), then the performance will probably just cap at this one-mover level.

TLDR: This commit is a short-term solution to a big problem.